Javascript file needed to create a post
This file is used to gather data from the create post page and then that
data is sent to the coldfusion Post.cfc to be inserted into the db.
Everything works as far as adding a post. There is still some
implementation to be done to make the create post UI work correctly.