Added code that allows a user to manually enter an expiration date
Added another table row to the UI to allow users to enter an expiration
date. Added a function that validates if the date entered is correct and
additional error messages if the date is not valid. Also, added another
argument to the Post.cfc that represents the date the user has entered.
This date is now used to set the expiration date on a post instead of
automatically setting the post to expire in 30 days.