Fixed Subscription/Notification problems
I added a javascript function that checks through the list of
subscriptions to see if a subscription already exist. If it does the
ajax call the user is alerted that they have already subscribed and the
ajax call is never made. Also fixed several issues in the Subscriber.cfc
and removed unneccessary code.